+++ CHANGES.md
@@ -5,6 +5,6 @@ Release Notes.
 Apollo 2.1.0
 
 ------------------
-
+* [Add a config adjust the property source overriden behavior](https://github.com/apolloconfig/apollo/pull/4409)
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/11?closed=1)